TS8: FIX_Оптимизировано тестирование: 1. Тесты из TaskTest, EpicTest и SubTaskTest вынесены в TaskManagerTest 2. В FileBackedTaskManagerTest реализовано удаление временных файлов после отработки тестов 3. В FileBackedTaskManager в методе main реализовано удаление создаваемого файла из папки sourses
