+++ sourses/tasks.csv
@@ -1,4 +1,4 @@
 id,type,name,status,description,startTime,duration,epic
-1,EPIC,Epic1,Новая,descr1,2024-03-10T19:48:21.841112100,PT9M
-2,SUBTASK,ST 1,Новая,descr1,10_03_2024 19:48,PT9M,1
+1,EPIC,Epic1,Новая,descr1,2024-03-10T21:16:31.899362900,PT9M
+2,SUBTASK,ST 1,Новая,descr1,10_03_2024 21:16,PT9M,1
